% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{iv_meta}
\alias{iv_meta}
\title{Inverse Variance Meta-Analysis}
\usage{
iv_meta(mean, sd, n, se = NULL)
}
\arguments{
\item{mean}{Vector of means}

\item{sd}{Vector of standard deviations}

\item{n}{Vector of sample sizes}

\item{se}{Vector of standard errors (optional)}
}
\value{
If se is NULL, returns weighted mean. Otherwise, returns a vector with weighted mean and confidence interval.
}
\description{
Inverse Variance Meta-Analysis
}
