% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_responder_analysis}
\alias{launch_responder_analysis}
\title{Launch Responder Analysis App}
\usage{
launch_responder_analysis()
}
\description{
This function launches the Responder Analysis Shiny application.
}
