% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{prop_meta}
\alias{prop_meta}
\title{Proportion Meta-Analysis}
\usage{
prop_meta(n_c = n_c, n_e = n_e, data)
}
\arguments{
\item{n_c}{Vector of event counts}

\item{n_e}{Vector of sample sizes}

\item{data}{Data frame containing n_c and n_e columns}
}
\value{
A vector with transformed proportion and variance
}
\description{
Proportion Meta-Analysis
}
