% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{rr_meta}
\alias{rr_meta}
\title{Risk Ratio Meta-Analysis}
\usage{
rr_meta(ai = ai, bi = bi, ci = ci, di = di, data)
}
\arguments{
\item{ai}{Vector of event counts in group 1}

\item{bi}{Vector of non-event counts in group 1}

\item{ci}{Vector of event counts in group 2}

\item{di}{Vector of non-event counts in group 2}

\item{data}{Data frame containing ai, bi, ci, di columns}
}
\value{
A vector with log risk ratio (multiplied by 100) and confidence interval
}
\description{
Risk Ratio Meta-Analysis
}
